requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.3
billiard==4.1.0
bleach==5.0.1
celery==5.3.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==39.0.2
defusedxml==0.7.1
dj-database-url==1.2.0
dj-email-url==1.0.6
Django==4.0.2
django-bleach==3.0.1
django-cache-url==3.4.4
django-ckeditor==6.5.1
django-editorjs==0.2.1
django-environ==0.11.2
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-rosetta==0.9.8
djangorestframework==3.14.0
environs==9.5.0
idna==3.4
jdatetime==4.1.1
kombu==5.3.2
marshmallow==3.19.0
oauthlib==3.2.2
packaging==23.0
Pillow==9.5.0
polib==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
rabbitmq==0.2.0
redis==4.6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
tzdata==2023.3
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
